fix: resolve all pyright and ruff errors across codebase
Add type annotations, assertions, casts, and type: ignore comments to
satisfy pyright standard mode. Fix ruff unused imports and dead code.
No logic changes — purely type-level fixes.

+++ src/wavern/core/audio_loader.py
@@ -108,7 +108,7 @@ class AudioLoader:
         import soundfile as sf
 
         audio, sr = sf.read(file_path, dtype="float32")
-        return audio, sr
+        return audio.astype(np.float32), sr
 
     @staticmethod
     def _load_pydub_fallback(file_path: str) -> tuple[NDArray[np.float32], int]:
